% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.key}
\alias{get.key}
\title{Pull the API key from \code{getOption('google.key')} or the
\code{GOOGLE_KEY} environment variable.}
\usage{
get.key()
}
\value{
The API key
}
\description{
Pull the API key from \code{getOption('google.key')} or the
\code{GOOGLE_KEY} environment variable.
}

