\name{languages}
\alias{languages}
\title{List the valid language mappings; optionally, from a given source
or to a given target.}
\usage{
  languages(source = NULL, target = NULL, key = get.key())
}
\arguments{
  \item{key}{The API key}

  \item{source}{The source language, e.g. \code{"en"}}

  \item{target}{The target language, e.g. \code{"de"}}
}
\description{
  List the valid language mappings; optionally, from a
  given source or to a given target.
}

