\name{translate}
\alias{translate}
\title{Translate the given text from a source to a target language.}
\usage{
  translate(query, source, target, key = get.key())
}
\arguments{
  \item{query}{The text to invoke against}

  \item{source}{The source language, e.g. \code{"en"}}

  \item{target}{The target language, e.g. \code{"de"}}

  \item{key}{The API key}
}
\description{
  Translate the given text from a source to a target
  language.
}

